Use a non-printable record delimiter over new line
By utilizing a this delimiter (U+001F - record separator https://en.wikipedia.org/wiki/Delimiter) we avoid delimiter collision. In this case, new lines in blobs of text would be interpretted as delimiter causing the bash array of key value pairs to be inaccurate. This leads to a bad param=value pair exception exposed as a DgiSaxonHelperTransformationException.